Samuel_Taiwo/requirements.txt
dlib==19.23.1
face_recognition==1.3.0
numpy==1.20.3
opencv_python==4.5.5.64
openface==0.0.0
Pillow==9.5.0
pytube==12.0.0
torch==1.11.0
